update chapter 6 recap

+++ Chapter6/dag_utils.R
@@ -1,5 +1,5 @@
 
-drawmydag = function(dag, cex = 1, lwd = 1) {
+drawmydag = function(dag, cex = .75, lwd = 1) {
   
   label_clr = rep("black",length(names(dag)))
   label_clr[which(names(dag) == latents(dag))] = "white"
@@ -43,7 +43,7 @@ get_xy = function(coord.list, nodes, axis, length.out = 200) {
                        coord.list[[axis]][x[2]],length.out = length.out))) 
 }
 
-plot_path = function(x,ys,k,arrow.col) {
+plot_path = function(x,ys,k, arrow.col = "red") {
   lines(x[1:k],ys[1:k],col = arrow.col, lwd = 2)
   arrows(
     x0 = x[k-1],
@@ -55,14 +55,18 @@ plot_path = function(x,ys,k,arrow.col) {
     length = .25/4
   )
 }
-flow_dag = function(dag, nodes, fn, arrow.col = "red", fps = 25, length.out = 200, nodes2 = NULL) {
+flow_dag = function(dag, nodes, fn = NULL, arrow.col = "red", arrow.col2 = NULL, cex = 1,
+                    fps = 25, length.out = 200, nodes2 = NULL, animation = TRUE, p.offset = 1.5) {
+  
+  arrow.col2 = ifelse(is.null(arrow.col2),arrow.col,arrow.col2)
+  
   coord.list = coordinates(dag)
   x = get_xy(coord.list, nodes, "x", length.out = length.out)
   y = -1*get_xy(coord.list, nodes, "y", length.out = length.out)
   
   ys = smth.gaussian(y, tails = F)
   ys[which(is.na(ys))] = y[which(is.na(ys))]
-  ys = ys + .05*1.5
+  ys = ys + (.05*p.offset)*ifelse(diff(coord.list$y[nodes[[1]]])>0,-1,1)
   
   if(!is.null(nodes2)) {
     lo = length.out/(length(nodes2)/length(nodes))
@@ -71,20 +75,26 @@ flow_dag = function(dag, nodes, fn, arrow.col = "red", fps = 25, length.out = 20
     
     ys2 = smth.gaussian(y2, tails = F)
     ys2[which(is.na(ys2))] = y2[which(is.na(ys2))]
-    ys2 = ys2 + .05*1.5
+    ys2 = ys2 + (.05*p.offset)*ifelse(diff(coord.list$y[nodes[[1]]])>0,-1,1)
   }
-  
-  file.remove(list.files("anim",full.names = T))
-  for (k in round(seq(2,length(x), length.out = 100))) {
-    png(paste0("anim/anim",sprintf("%03d", k), ".png"),
-        width = 3, height = 2.5, units = "in", res = 300)
-    drawmydag(dag, cex = 1)
-    plot_path(x,ys,k,arrow.col = arrow.col)
+  if (animation == TRUE) {
+    file.remove(list.files("anim",full.names = T))
+    for (k in round(seq(2,length(x), length.out = 100))) {
+      png(paste0("anim/anim",sprintf("%03d", k), ".png"),
+          width = 3, height = 2.5, units = "in", res = 300)
+      drawmydag(dag, cex = cex)
+      plot_path(x,ys,k,arrow.col = arrow.col)
+      if (!is.null(nodes2)) 
+        plot_path(x2,ys2,k,arrow.col = arrow.col2)
+      dev.off()
+    }
+    make_gif("anim",fn, fps = fps)
+  } else {
+    drawmydag(dag, cex = cex)
+    plot_path(x,ys,length(x),arrow.col = arrow.col)
     if (!is.null(nodes2)) 
-      plot_path(x2,ys2,k,arrow.col = arrow.col)
-    dev.off()
+      plot_path(x2,ys2,length(x),arrow.col = arrow.col2)
   }
-  make_gif("anim",fn, fps = 25)
 }
 
 make_gif = function(dir = NULL, fn = NULL, fps = 20) {
